Abstract common methods to a base logger class
