+++ php/WP_CLI/Loggers/Quiet.php
@@ -5,7 +5,7 @@ namespace WP_CLI\Loggers;
 /**
  * Quiet logger only logs errors.
  */
-class Quiet {
+class Quiet extends Base {
 
 	/**
 	 * Informational messages aren't logged.
@@ -26,18 +26,6 @@ class Quiet {
 	}
 
 	/**
-	 * Write a message to STDERR, prefixed with "Debug: ".
-	 *
-	 * @param string $message Message to write.
-	 */
-	public function debug( $message ) {
-		if ( \WP_CLI::get_runner()->config['debug'] ) {
-			$time = round( microtime( true ) - WP_CLI_START_MICROTIME, 3 );
-			$this->_line( "$message ({$time}s)", 'Debug', '%B', STDERR );
-		}
-	}
-
-	/**
 	 * Warning messages aren't logged.
 	 *
 	 * @param string $message Message to write.
@@ -52,7 +40,7 @@ class Quiet {
 	 * @param string $message Message to write.
 	 */
 	public function error( $message ) {
-		fwrite( STDERR, \WP_CLI::colorize( "%RError:%n $message\n" ) );
+		$this->write( STDERR, \WP_CLI::colorize( "%RError:%n $message\n" ) );
 	}
 
 	/**
@@ -63,7 +51,7 @@ class Quiet {
 	public function error_multi_line( $message_lines ) {
 		$message = implode( "\n", $message_lines );
 
-		fwrite( STDERR, \WP_CLI::colorize( "%RError:%n\n$message\n" ) );
-		fwrite( STDERR, \WP_CLI::colorize( "%R---------%n\n\n" ) );
+		$this->write( STDERR, \WP_CLI::colorize( "%RError:%n\n$message\n" ) );
+		$this->write( STDERR, \WP_CLI::colorize( "%R---------%n\n\n" ) );
 	}
 }
